requirements.txt
streamlit
GoogleNews
pandas
sqlite-utils
dateparser
groq
requests
python-dotenv
openai
pytz
firebase-admin
